Restart systemd.resolv only if not resolv.conf exists

+++ files/firstboot/25-create-resolv-symlink.sh
@@ -1,7 +1,7 @@
 logger -t "rc.firstboot" "Creating /etc/resolv.conf symlink"
 
 # Check if systemd resolve directory exists
-if [ ! -d "/run/systemd/resolve" ] ; then
+if [ ! -d "/run/systemd/resolve" -a ! -e "/etc/resolv.conf" ] ; then
   systemctl enable systemd-resolved.service
   systemctl restart systemd-resolved.service
 fi
@@ -11,5 +11,8 @@ if [ ! -f "/run/systemd/resolve/resolv.conf" ] ; then
   touch /run/systemd/resolve/resolv.conf
 fi
 
-# Create symlink to /etc/reolv.conf
-ln -s /run/systemd/resolve/resolv.conf /etc/resolv.conf
+# Create symlink to /etc/reolv.conf if not exists yet
+if [ ! -e "/etc/resolv.conf" ] ; then
+  ln -s /run/systemd/resolve/resolv.conf /etc/resolv.conf
+fi
+
